chore: addressed feedback from gemini
